Renamed SimpleMetadataService to ConformanceMetadataService since that's now its only purpose
